+++ tests/testthat/test-workStudy.R
@@ -175,24 +175,28 @@ test_that('Test getDetectorData', {
 test_that('Test updateFiles', {
     data(exStudy)
     # corrupting filepaths
-    recs <- system.file('extdata', 'Recordings', package='PAMpal')
-    exStudy <- addRecordings(exStudy, folder =recs, log=FALSE, progress=FALSE)
     files(exStudy)$db <- substr(files(exStudy)$db, start=5, stop=10e3)
     files(exStudy)$binaries <- substr(files(exStudy)$binaries, start=5, stop=10e3)
-    files(exStudy)$recordings$file <- substr(files(exStudy)$recordings$file, start=5, stop=10e3)
     files(exStudy[[1]])$db <- substr(files(exStudy[[1]])$db, start=5, stop=10e3)
     files(exStudy[[1]])$binaries <- substr(files(exStudy[[1]])$binaries, start=5, stop=10e3)
     db <- system.file('extdata', 'Example.sqlite3', package='PAMpal')
     bin <- system.file('extdata', 'Binaries', package='PAMpal')
     expect_true(!any(file.exists(files(exStudy)$db,
                                  files(exStudy)$binaries,
-                                 files(exStudy)$recordings$file,
+                                 # files(exStudy)$recordings$file,
                                  files(exStudy[[1]])$db,
                                  files(exStudy[[1]])$binaries)))
-    exStudy <- updateFiles(exStudy, db=db, bin=bin, recording = recs, verbose=FALSE)
+    exStudy <- updateFiles(exStudy, db=db, bin=bin, verbose=FALSE)
+    # exStudy <- updateFiles(exStudy, db=db, bin=bin, recording = recs, verbose=FALSE)
     expect_true(all(file.exists(files(exStudy)$db,
                                 files(exStudy)$binaries,
-                                files(exStudy)$recordings$file,
+                                # files(exStudy)$recordings$file,
                                 files(exStudy[[1]])$db,
                                 files(exStudy[[1]])$binaries)))
+    recs <- system.file('extdata', 'Recordings', package='PAMpal')
+    exStudy <- addRecordings(exStudy, folder =recs, log=FALSE, progress=FALSE)
+    files(exStudy)$recordings$file <- substr(files(exStudy)$recordings$file, start=5, stop=10e3)
+    expect_true(!any(file.exists(files(exStudy)$recordings$file)))
+    exStudy <- updateFiles(exStudy, recording=recs, verbose=FALSE)
+    expect_true(all(file.exists(files(exStudy)$recordings$file)))
 })